feature_33 start to remove BD from MainActivity
