+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.AndroidPraktika"
         tools:targetApi="31">
         <activity
-            android:name=".SplashScreenActivity"
+            android:name=".view.SplashScreenActivity"
             android:exported="true"
             android:theme="@style/Theme.AndroidPraktika.NoActionBar">
             <intent-filter>
@@ -27,7 +27,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:exported="false"
             android:theme="@style/Theme.AndroidPraktika.NoActionBar">
