--- app/src/main/java/ed/maevski/androidpraktika/data/Ad.kt
@@ -1,3 +1,0 @@
-package ed.maevski.androidpraktika.data
-
-class Ad(override val id: Int, val title: String, val content : String, override var isInFavorites: Boolean = false) : Item
