+++ app/src/main/java/ed/maevski/androidpraktika/data/MainRepository.kt
@@ -1,16 +1,11 @@
-package ed.maevski.androidpraktika
+package ed.maevski.androidpraktika.data
 
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import ed.maevski.androidpraktika.data.Ad
-import ed.maevski.androidpraktika.data.DeviantPicture
-import ed.maevski.androidpraktika.databinding.ActivityMainBinding
+import ed.maevski.androidpraktika.R
+import ed.maevski.androidpraktika.domain.Ad
+import ed.maevski.androidpraktika.domain.DeviantPicture
 
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-
-    private val itemsRV = listOf(
+class MainRepository {
+    val itemsRV = listOf(
         DeviantPicture(
             0,
             "Casual Mulan",
@@ -123,84 +118,4 @@ class MainActivity : AppCompatActivity() {
             65000
         )
     )
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        supportFragmentManager
-            .beginTransaction()
-            .add(R.id.fragment_placeholder, HomeFragment(itemsRV))
-            .addToBackStack(null)
-            .commit()
-
-        initMenu()
-    }
-
-    //Ищем фрагмент по тегу, если он есть то возвращаем его, если нет, то null
-    private fun checkFragmentExistence(tag: String): Fragment? =
-        supportFragmentManager.findFragmentByTag(tag)
-
-    private fun changeFragment(fragment: Fragment, tag: String) {
-        supportFragmentManager
-            .beginTransaction()
-            .replace(R.id.fragment_placeholder, fragment, tag)
-            .addToBackStack(null)
-            .commit()
-    }
-
-    fun launchDetailsFragment(picture: DeviantPicture) {
-        //Создаем "посылку"
-        val bundle = Bundle()
-        //Кладем в "посылку"
-        bundle.putParcelable("dev", picture)
-        //Кладем фрагмент с деталями в перменную
-        val fragment = DetailsFragment()
-        //Прикрепляем нашу "посылку" к фрагменту
-        fragment.arguments = bundle
-
-        //Запускаем фрагмент
-        supportFragmentManager
-            .beginTransaction()
-            .replace(R.id.fragment_placeholder, fragment)
-            .addToBackStack(null)
-            .commit()
-    }
-
-    private fun initMenu() {
-        binding.bottomNavigation.setOnItemSelectedListener {
-
-            when (it.itemId) {
-                R.id.home -> {
-                    val tag = "home"
-                    val fragment = checkFragmentExistence(tag)
-                    //В первом параметре, если фрагмент не найден и метод вернул null, то с помощью
-                    //элвиса мы вызываем создание нового фрагмента
-                    changeFragment(fragment ?: HomeFragment(itemsRV), tag)
-                    true
-                }
-                R.id.favorites -> {
-                    val tag = "favorites"
-                    val fragment = checkFragmentExistence(tag)
-                    changeFragment(fragment ?: FavoritesFragment(itemsRV), tag)
-                    true
-                }
-                R.id.watch_later -> {
-                    val tag = "watch_later"
-                    val fragment = checkFragmentExistence(tag)
-                    changeFragment(fragment ?: WatchLaterFragment(), tag)
-                    true
-                }
-                R.id.selections -> {
-                    val tag = "selections"
-                    val fragment = checkFragmentExistence(tag)
-                    changeFragment(fragment ?: SelectionsFragment(), tag)
-                    true
-                }
-                else -> false
-            }
-        }
-    }
 }
