+++ app/src/main/java/ed/maevski/androidpraktika/domain/Ad.kt
@@ -1,0 +1,4 @@
+package ed.maevski.androidpraktika.domain
+
+class Ad(override val id: Int, val title: String, val content : String, override var isInFavorites: Boolean = false) :
+    Item
