+++ app/src/main/java/ed/maevski/androidpraktika/domain/Interactor.kt
@@ -1,0 +1,7 @@
+package ed.maevski.androidpraktika.domain
+
+import ed.maevski.androidpraktika.data.MainRepository
+
+class Interactor(val repo: MainRepository) {
+    fun getPicturesDB(): List<Item> = repo.itemsRV
+}
