+++ app/src/main/java/ed/maevski/androidpraktika/view/adapters/PictureDelegateAdapter.kt
@@ -1,13 +1,14 @@
-package ed.maevski.androidpraktika.adapter
+package ed.maevski.androidpraktika.view.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.hannesdorfmann.adapterdelegates4.AbsListItemAdapterDelegate
-import ed.maevski.androidpraktika.data.DeviantPicture
-import ed.maevski.androidpraktika.data.Item
+import ed.maevski.androidpraktika.domain.DeviantPicture
+import ed.maevski.androidpraktika.domain.Item
 import ed.maevski.androidpraktika.databinding.ItemPictureBinding
+import ed.maevski.androidpraktika.view.rv_adapters.PictureRecyclerAdapter
 
 class PictureDelegateAdapter(private val clickListener: PictureRecyclerAdapter.OnItemClickListener) :
     AbsListItemAdapterDelegate<DeviantPicture, Item, PictureDelegateAdapter.ViewHolder>() {
