+++ app/src/main/java/ed/maevski/androidpraktika/view/fragments/DetailsFragment.kt
@@ -1,4 +1,4 @@
-package ed.maevski.androidpraktika
+package ed.maevski.androidpraktika.view.fragments
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,7 +6,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import ed.maevski.androidpraktika.data.DeviantPicture
+import ed.maevski.androidpraktika.R
+import ed.maevski.androidpraktika.domain.DeviantPicture
 import ed.maevski.androidpraktika.databinding.FragmentDetailsBinding
 
 class DetailsFragment : Fragment() {
