+++ app/src/main/java/ed/maevski/androidpraktika/view/fragments/SelectionsFragment.kt
@@ -1,4 +1,4 @@
-package ed.maevski.androidpraktika
+package ed.maevski.androidpraktika.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,24 +6,29 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.activity.OnBackPressedCallback
-import ed.maevski.androidpraktika.databinding.FragmentWatchLaterBinding
+import ed.maevski.androidpraktika.databinding.FragmentSelectionsBinding
+import ed.maevski.androidpraktika.utils.AnimationHelper
 
-class WatchLaterFragment : Fragment() {
-    private var _binding: FragmentWatchLaterBinding? = null
+class SelectionsFragment : Fragment() {
+    private var _binding: FragmentSelectionsBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentWatchLaterBinding.inflate(inflater, container, false)
+        _binding = FragmentSelectionsBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        AnimationHelper.performFragmentCircularRevealAnimation(binding.watchLaterFragmentRoot, requireActivity(), 1)
+        AnimationHelper.performFragmentCircularRevealAnimation(
+            binding.selectionsFragmentRoot,
+            requireActivity(),
+            1
+        )
 
         requireActivity().onBackPressedDispatcher.addCallback(
             viewLifecycleOwner,
@@ -35,7 +40,7 @@ class WatchLaterFragment : Fragment() {
 
     override fun onDestroyView() {
         _binding = null
+
         super.onDestroyView()
     }
 }
-
