+++ app/src/main/java/ed/maevski/androidpraktika/view/rv_adapters/FavoriteRecyclerAdapter.kt
@@ -1,9 +1,11 @@
-package ed.maevski.androidpraktika.adapter
+package ed.maevski.androidpraktika.view.rv_adapters
 
 import com.hannesdorfmann.adapterdelegates4.ListDelegationAdapter
-import ed.maevski.androidpraktika.data.Item
+import ed.maevski.androidpraktika.domain.Item
+import ed.maevski.androidpraktika.view.adapters.AdDelegateAdapter
+import ed.maevski.androidpraktika.view.adapters.FavoriteDelegateAdapter
 
-class FavoriteRecyclerAfapter() : ListDelegationAdapter<List<Item>>() {
+class FavoriteRecyclerAdapter() : ListDelegationAdapter<List<Item>>() {
 
     init {
         delegatesManager.addDelegate(AdDelegateAdapter())
