+++ app/src/main/java/ed/maevski/androidpraktika/viewmodel/MainActivityViewModel.kt
@@ -1,0 +1,17 @@
+package ed.maevski.androidpraktika.viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import ed.maevski.androidpraktika.domain.Interactor
+import ed.maevski.androidpraktika.domain.Item
+
+class MainActivityViewModel : ViewModel() {
+    val picturesListLiveData = MutableLiveData<List<Item>>()
+
+    private lateinit var interactor: Interactor
+
+    init {
+        val films = interactor.getPicturesDB()
+        picturesListLiveData.postValue(films)
+    }
+}
