+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:id="@+id/splashScreenRoot"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SplashScreenActivity">
+    tools:context=".view.SplashScreenActivity">
 
     <com.airbnb.lottie.LottieAnimationView
         android:id="@+id/lottie_anim"
