+++ app/src/main/res/layout/fragment_selections.xml
@@ -7,6 +7,6 @@
     android:visibility="invisible"
     tools:visibility="visible"
     android:background="@drawable/background_selections"
-    tools:context=".SelectionsFragment">
+    tools:context=".view.fragments.SelectionsFragment">
 
 </FrameLayout>
